+++ Juego-Apocaba-master/CHANGELOG.md
@@ -1,3 +1,6 @@
-## [1.0.0] - 2026-08-25
+## [1.0.1] - 2026-09-14
 ### Added
 -- creacion del proyecto
+--creacion del jugador
+--movimiento horizontal y vertical mas gravedad
+--crecion de fondo